% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TimeAx.R
\name{detectTrajGenes}
\alias{detectTrajGenes}
\title{Selecting conserved-dynamics-seed features}
\usage{
detectTrajGenes(
  GEData,
  sampleNames,
  numOfTopGenes = 50,
  topGenes = 4000,
  numOfIterations = 20,
  percOfSamples = 0.8,
  no_cores = NULL
)
}
\arguments{
\item{GEData}{A matrix containing profiles (columns) of omics measurments (rows) from multiple individuals and different time points. Profiles for each individual should be ordered by chronological time.}

\item{sampleNames}{A vector containing the individual identity of each sample in the GEData.}

\item{numOfTopGenes}{Length of the conserved-dynamics-seed of features. The default is 50.}

\item{topGenes}{Number of initial high variable features to be considered for the seed selection. The default is 4000.}

\item{numOfIterations}{Number of different random sample selections for the calculation. The default is 20.}

\item{percOfSamples}{Fraction of samples from each individual, selected in each sample selection. The default is 0.8}

\item{no_cores}{A number for the amount of cores which will be used for the analysis. The default (NULL) is total number of cores minus 1.}
}
\value{
A list including:
The conserved-dynamics-seed. A list of feature, suitable for the model training, ordered from best to worst.
}
\description{
Selecting conserved-dynamics-seed features
}
\examples{
data(UBCData)

# Selecting conserved-dynamics-seed features
seed = detectTrajGenes(DataUBC,UBCSamples, no_cores = 2)
}
\references{
Submitted
}
