% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TimeAx.R
\name{modelCreation}
\alias{modelCreation}
\title{Training the TimeAx model}
\usage{
modelCreation(
  trainData,
  sampleNames,
  ratio = T,
  numOfIter = 100,
  numOfTopFeatures = 50,
  seed = NULL,
  no_cores = NULL
)
}
\arguments{
\item{trainData}{A matrix containing profiles (columns) of omics measurments (rows) from multiple individuals and different time points. For omics data it is better to use raw values instead of normalized ones. Profiles for each individual should be ordered by chronological time.}

\item{sampleNames}{A vector containing the individual identity of each sample in the train data.}

\item{ratio}{Boolean parameter determining whether the model will be based on the ratios between feature values or the raw values. The default is to use ratios (TRUE) as this allows the model to overcome technical batch effects. For data types with many zeros or discrete values, the raw data should be used (selecting FALSE).}

\item{numOfIter}{Number of consensus trajectories. The default is 100.}

\item{numOfTopFeatures}{Length of the conserved-dynamics-seed of features. If trainData has less features, all of them will be selected as seed. The default is 50.}

\item{seed}{The conserved-dynamics-seed. If provided, the alignment process will be conducted based on these features. The default is NULL.}

\item{no_cores}{A number for the amount of cores which will be used for the analysis. The defalt (NULL) is total number of cores minus 1.}
}
\value{
A TimeAx model consists of:
\item{consensusList}{List of consensus trajectories.}
\item{seed}{The conserved-dynamics-seed that was used in the alignment process.}
}
\description{
This function initiate model training using the TimeAx algorithm - performing a multiple trajectory alignment (MTA) on time-series datasets of individuals each of which is considered as an individual partial trajectory.
}
\examples{
data(UBCData)

# Training the model
model = modelCreation(DataUBC,UBCSamples, no_cores = 2)

\dontrun{

}
}
\references{
Submitted
}
