% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TimeAx.R
\name{predictByConsensus}
\alias{predictByConsensus}
\title{Infer pseudotime for new samples bassed on the TimeAx model}
\usage{
predictByConsensus(
  model,
  testData,
  no_cores = NULL,
  seed = NULL,
  sampleNames = NULL
)
}
\arguments{
\item{model}{A TimeAx model.}

\item{testData}{A matrix containing profiles (columns) of features measurments (rows). Data should provided in similar scales (preferably, non-normalized) as the train data.}

\item{no_cores}{A number for the amount of cores which will be used for the analysis. The default (NULL) is total number of cores minus 1.}

\item{seed}{The conserved-dynamics-seed. If provided, the prediction process will be conducted based on these features. Use the model's seed by keeping the the default value of NULL.}

\item{sampleNames}{Used for the robustness analysis. Always keep as NULL.}
}
\value{
A prediction list consists of:
\item{predictions}{The final pseudotime position for each sample.}
\item{uncertainty}{An uncertainty score for each position. Lower scores means higher certainty.}
}
\description{
Infer pseudotime for new samples bassed on the TimeAx model
}
\examples{
data(UBCData)

# Training the model
model = modelCreation(DataUBC,UBCSamples, no_cores = 2)

# Inferring pseudotime positions
pseudotimeStats = predictByConsensus(model,DataUBC, no_cores = 2)
pseudotime = pseudotimeStats$predictions
uncertainty = pseudotimeStats$uncertainty
}
\references{
Submitted
}
