% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TimeAx.R
\name{robustness}
\alias{robustness}
\title{Calculate a robustness score for the TimeAx model}
\usage{
robustness(model, trainData, sampleNames, pseudo = NULL, no_cores = NULL)
}
\arguments{
\item{model}{A TimeAx model.}

\item{trainData}{The matrix containing profiles (columns) of omics measurments (rows), which was used to train the model.}

\item{sampleNames}{A vector containing the individual identity of each sample in the train data. Same vector as used in the training.}

\item{pseudo}{The output list of predictByConsensus. If not provided (NULL), pseudotime will be inferred by this function.}

\item{no_cores}{A number for the amount of cores which will be used for the analysis. The default (NULL) is total number of cores minus 1.}
}
\value{
A robustness list consists of:
\item{robustnessPseudo}{Robustness pseudotime positions for all samples.}
\item{score}{TimeAx robustness score for the model.}
}
\description{
Calculate a robustness score for the TimeAx model
}
\examples{
data(UBCData)

# Training the model
model = modelCreation(DataUBC,UBCSamples,no_cores = 2)

# Inferring pseudotime positions
robustnessStats = robustness(model,DataUBC,UBCSamples,no_cores = 2)
robustnessPseudo = robustnessStats$robustnessPseudo
robustnessScore = robustnessStats$score
}
\references{
Submitted
}
