% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TimeAx.R
\docType{data}
\name{UBCSamples}
\alias{UBCSamples}
\title{UBC sample labels.}
\format{
A vector with 27 sample labels for 5 individuals.
}
\usage{
UBCSamples
}
\description{
UBC sample labels.
}
\keyword{datasets}
