% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TimeAx.R
\name{modelCreation}
\alias{modelCreation}
\title{Training the TimeAx model}
\usage{
modelCreation(
  GEData,
  sampleNames,
  numOfIter = 100,
  numOfTopGenes = 50,
  seed = NULL,
  no_cores = NULL
)
}
\arguments{
\item{GEData}{A matrix containing profiles (columns) of omics measurments (rows) from multiple individuals and different time points. Profiles for each individual should be ordered by chronological time.}

\item{sampleNames}{A vector containing the individual identity of each sample in the GEData.}

\item{numOfIter}{Number of consensus trajectories. The default is 100.}

\item{numOfTopGenes}{Length of the conserved-dynamics-seed of features. The default is 50.}

\item{seed}{The conserved-dynamics-seed. If provided, the alignment process will be conducted based on these features. The default is NULL.}

\item{no_cores}{A number for the amount of cores which will be used for the analysis. The defalt (NULL) is total number of cores minus 1.}
}
\value{
A TimeAx model consists of:
\item{consensusList}{List of consensus trajectories.}
\item{seed}{The conserved-dynamics-seed that was used in the alignment process.}
}
\description{
This function initiate model training using the TimeAx algorithm - performing a multiple trajectory alignment (MTA) on time-series datasets of individuals each of which is considered as an individual partial trajectory.
}
\examples{
data(UBCData)

# Training the model
model = modelCreation(DataUBC,UBCSamples, no_cores = 2)

\dontrun{

}
}
\references{
Submitted
}
